Fix start game, pre start game, and start act hooks.
