--- src/main/java/basemod/patches/com/megacrit/cardcrawl/cards/AbstractCard/RenderEnergySwitch.java
@@ -1,78 +1,0 @@
-package basemod.patches.com.megacrit.cardcrawl.cards.AbstractCard;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.evacipated.cardcrawl.modthespire.lib.SpireInsertPatch;
-import com.evacipated.cardcrawl.modthespire.lib.SpirePatch;
-import com.megacrit.cardcrawl.cards.AbstractCard;
-import com.megacrit.cardcrawl.cards.AbstractCard.CardColor;
-import com.megacrit.cardcrawl.helpers.ImageMaster;
-
-import basemod.BaseMod;
-import basemod.abstracts.CustomCardWithRender;
-import basemod.helpers.SuperclassFinder;
-
-@SpirePatch(cls = "com.megacrit.cardcrawl.cards.AbstractCard", method = "renderEnergy")
-public class RenderEnergySwitch {
-	public static final Logger logger = LogManager.getLogger(BaseMod.class.getName());
-	
-	@SpireInsertPatch(rloc = 26, localvars = { "drawX", "drawY" })
-	public static void Insert(Object __obj_instance, Object sbObj, float drawX, float drawY) {
-		AbstractCard card = (AbstractCard) __obj_instance;
-		CardColor color = card.color;
-		SpriteBatch sb = (SpriteBatch) sbObj;
-		
-		if (!color.toString().equals("RED") && !color.toString().equals("GREEN") && !color.toString().equals("BLUE")
-				&& !color.toString().equals("COLORLESS") && !color.toString().equals("CURSE")) {
-			Texture orbTexture = null;
-			try {
-				if (card instanceof CustomCardWithRender) {
-					orbTexture = ((CustomCardWithRender)card).getOrbSmallTexture();
-				} 
-				if (orbTexture == null){
-					Texture baseModTexture = BaseMod.getEnergyOrbTexture(color.toString());
-					if (baseModTexture == null) {
-						orbTexture = new Texture(BaseMod.getEnergyOrb(color.toString()));
-						BaseMod.saveEnergyOrbTexture(color.toString(), orbTexture);
-					} else {
-						orbTexture = baseModTexture;
-					}
-				}
-			} catch (NullPointerException e) {
-				logger.error("could not load texture for energy orb for card " + card.getClass().toString() + " with color " + color.toString());
-				logger.error("exception is: " + e.getMessage());
-				e.printStackTrace();
-				orbTexture = ImageMaster.CARD_COLORLESS_ORB;
-			}
-			try {
-				// use reflection hacks to invoke renderHelper (without float scale)
-				Method renderHelperMethod;
-				Field renderColorField; 
-				
-				
-				renderHelperMethod = SuperclassFinder.getSuperClassMethod(card.getClass(), "renderHelper", SpriteBatch.class, Color.class, Texture.class, float.class, float.class);
-				renderHelperMethod.setAccessible(true);
-				renderColorField = SuperclassFinder.getSuperclassField(card.getClass(), "renderColor");
-				renderColorField.setAccessible(true);
-				
-					
-				Color renderColor = (Color) renderColorField.get(card);
-				renderHelperMethod.invoke(card, sb, renderColor, orbTexture, drawX, drawY);
-			} catch (IllegalAccessException | IllegalArgumentException | NoSuchFieldException | NoSuchMethodException | InvocationTargetException | SecurityException e) {
-				logger.error("could not render energy for card " + card.getClass().toString() + " with color " + color.toString());
-				logger.error("exception is: " + e.getMessage());
-				e.printStackTrace();
-			}
-
-		}
-
-	}
-}
