--- src/main/java/basemod/patches/com/megacrit/cardcrawl/core/CardCrawlGame/PostUpdateHook.java
@@ -1,30 +1,0 @@
-package basemod.patches.com.megacrit.cardcrawl.core.CardCrawlGame;
-
-import basemod.BaseMod;
-import com.evacipated.cardcrawl.modthespire.lib.*;
-import javassist.CannotCompileException;
-import javassist.CtBehavior;
-
-import java.util.ArrayList;
-
-import com.evacipated.cardcrawl.modthespire.patcher.PatchingException;
-import com.megacrit.cardcrawl.helpers.InputHelper;
-
-@SpirePatch(cls="com.megacrit.cardcrawl.core.CardCrawlGame", method="update")
-public class PostUpdateHook {
-	
-    @SpireInsertPatch
-    public static void Insert(Object __obj_instance) {
-        BaseMod.publishPostUpdate();
-    }
-
-    public static class Locator extends SpireInsertLocator
-    {
-        public int[] Locate(CtBehavior ctMethodToPatch) throws CannotCompileException, PatchingException
-        {
-            Matcher finalMatcher = new Matcher.MethodCallMatcher(InputHelper.class.getName(), "updateLast");
-
-            return LineFinder.findInOrder(ctMethodToPatch, new ArrayList<Matcher>(), finalMatcher);
-        }
-    }
-}
